+++ src/userinfo.jl
@@ -41,8 +41,8 @@ function _get_authenticated_user_legacy(server::AbstractString, token::Secret)::
     r = Mocking.@mock _get_authenticated_user_legacy_gql_request(server, token)
     msg = "Unable to query for user information (Hasura)" # error message
     r.status == 200 || _throw_invalidresponse(r; msg)
-    json, _ = _parse_response_json(r, Dict)
-    users = _get_json(_get_json(json, "data", Dict; msg), "users", Vector)
+    json, _ = _parse_response_json(r, AbstractDict)
+    users = _get_json(_get_json(json, "data", AbstractDict; msg), "users", Vector)
     length(users) == 1 || throw(
         JuliaHubError("$msg\nInvalid JSON returned by the server: length(users)=$(length(users))")
     )
@@ -60,7 +60,7 @@ function _get_api_information(server::AbstractString, token::Secret)::_JuliaHubI
     # First, try to access the /api/v1 endpoint
     r = Mocking.@mock _get_authenticated_user_api_v1_request(server, token)
     if r.status == 200
-        json, _ = _parse_response_json(r, Dict)
+        json, _ = _parse_response_json(r, AbstractDict)
         username = _get_json_or(json, "username", String, nothing)
         api_version = _json_get(json, "api_version", VersionNumber; parse=true, var="/api/v1")
         return _JuliaHubInfo(; username, api_version)